+++ doc/changelog.rst
@@ -3,6 +3,7 @@
 Release Notes
 =============
 
+- :feature:`-` You can now exclude proposals with pending state changes from your search results.
 - :feature:`-` To improve email template handling, the list of emails now shows just the subjects or use case, and you can click them to expand and see the details.
 - :feature:`-` Breaks are now also shown on the mobile/minimal/linear schedule.
 - :bug:`-` Review pages were not working when pretalx was run with Python 3.7 and the aggregation method "mean" (as opposed to "median").